Add account statement api

+++ src/utils/api/apiService.ts
@@ -85,9 +85,11 @@ const apiHandler: (arg: ApiServiceInterface) => Promise<ApiResponse> = async (
       result["error"] = error.response?.data;
       console.log(error.response.status)
       if (error.response?.status === 401) {
-        localStorage.clear();
-        snackBarUtil.error("Session changed. Please login again!");
-        window.location.replace("/");
+        if (localStorage.getItem('token')) {
+          localStorage.clear();
+          snackBarUtil.error("Session changed. Please login again!");
+          window.location.replace("/");
+        }
       }
     })
     .then((response) => {
@@ -114,11 +116,24 @@ const apiSnackbarNotifications: (
   }
   return args;
 };
-
+const apiWithErrorSnackbar: (
+  arg: ApiServiceInterface
+) => Promise<ApiResponse> = async (args) => {
+  const result = await apiHandler(args);
+  if (result?.error) {
+    const { message } = result.error;
+    if (typeof message === "object") {
+      message?.forEach((message) => snackBarUtil.error(message));
+    } else {
+      snackBarUtil.error(message);
+    }
+  }
+  return result;
+};
 const apiWithSnackbar: (
   arg: ApiServiceInterface
 ) => Promise<ApiResponse> = async (args) => {
   const result = await apiHandler(args);
   return apiSnackbarNotifications(result);
 };
-export { apiService, apiHandler, apiWithSnackbar };
+export { apiService, apiHandler, apiWithErrorSnackbar, apiWithSnackbar };
